@@ -1,21 +1,21 @@
 import 'package:check_feng_shui/helper/i18n.dart';
 import 'package:flutter/material.dart';
 
-class DefaultLocalizations extends LocalizationsDelegate<I18n> {
-  const DefaultLocalizations();
+// class DefaultLocalizations extends LocalizationsDelegate<I18n> {
+//   const DefaultLocalizations();
 
-  @override
-  bool isSupported(Locale locale) {
-    return ['en', 'vi'].contains(locale.languageCode);
-  }
+//   @override
+//   bool isSupported(Locale locale) {
+//     return ['en', 'vi'].contains(locale.languageCode);
+//   }
 
-  @override
-  Future<I18n> load(Locale locale) {
-    return I18n.load(locale);
-  }
+//   @override
+//   Future<I18n> load(Locale locale) {
+//     return I18n.load(locale);
+//   }
 
-  @override
-  bool shouldReload(DefaultLocalizations old) {
-    return false;
-  }
-}
+//   @override
+//   bool shouldReload(DefaultLocalizations old) {
+//     return false;
+//   }
+// }
