@@ -23,10 +23,8 @@ class MyApp extends StatelessWidget {
 }
 
 class MyHomePage extends StatelessWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
+  MyHomePage({required this.title});
+  String title;
   @override
   Widget build(BuildContext context) {
     return MainScreenScaffold();
